services/pipeline-runner/requirements.txt
grpcio==1.70.0

fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0

google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0

requests==2.32.3
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0
